UDP error handling: Only fire callback once
The error handler for UDP sockets can fire on socket.send(), for example if
DNS resolution fails temporarily. While we do persistently register an error
handler, it is not a good idea to repeatedly call the error handler. The
callback is expected to only be called once, and callers are likely to fail in
entertaining ways if this expectation is violated.

To enforce call-once semantics, this patch just sets the local `done` callback
to null after calling it once. Subsequent `error` callbacks won't call it
again.